+++ src/app/olmap/olmap.js
@@ -3,15 +3,16 @@
 
 angular.module('odssPlatimApp.olmap', ['odssPlatimApp.olmap.directives'])
     .controller('MapCtrl', MapCtrl)
-    .factory('olmap', olmapFactory)
+    .factory('olMap', olMap)
 ;
 
-MapCtrl.$inject = ['$scope', 'olmap'];
+MapCtrl.$inject = ['$scope', 'olMap'];
 
-function MapCtrl($scope, olmap) {
-    olmap.insertMap();
+function MapCtrl($scope, olMap) {
+    olMap.insertMap();
 
-    $scope.mv = {
+    var vm = {
+        viewOnly: true,
         editMode: false,
         draw: {
             typeList: [
@@ -23,38 +24,60 @@ function MapCtrl($scope, olmap) {
             selectedType: ""
         }
     };
-    $scope.$watch('mv.draw.selectedType', function(type) {
-        olmap.setDrawType(type);
-    });
-    $scope.$watch('mv.editMode', function(editMode) {
-        olmap.setEditMode(editMode);
-        if (editMode) {
-            olmap.setDrawType($scope.mv.draw.selectedType);
+    $scope.vm = vm;
+    $scope.$watch('vm.editMode', function(editMode) {
+        olMap.setEditMode(editMode);
+        if (!editMode) {
+            vm.draw.selectedType = "";
         }
+        olMap.setDrawType(vm.draw.selectedType);
+    });
+    $scope.$watch('vm.draw.selectedType', olMap.setDrawType);
+
+    $scope.$on("tokenSelection", function(e, selected) {
+        vm.viewOnly = selected.length != 1;
+        if (vm.viewOnly) {
+            olMap.setEditMode(false);
+            vm.editMode = false;
+            vm.draw.selectedType = "";
+            olMap.setDrawType(vm.draw.selectedType);
+        }
+        olMap.setTokenSelection(selected);
     });
 }
 
-olmapFactory.$inject = [];
+olMap.$inject = ['$rootScope'];
 
-function olmapFactory() {
+function olMap($rootScope) {
+    var styles = getStyles();
 
-    var gmap = createGMap();
-    var view = createView();
-    var layers = createLayers();
+    var gmap, view, vectorLayer;
     var map, draw, modify, featureOverlay;
+    var vectorsByGeomId = {};
+
+    var tokenSelection = [];
+    var editInfo = {
+        editingToken: null
+    };
 
     return {
-        insertMap:   insertMap
-        ,setCenter:  setCenter
-        ,setZoom:    setZoom
-        ,setEditMode:  setEditMode
-        ,setDrawType:  setDrawType
+        insertMap:           insertMap
+        ,reinit:             reinit
+        ,addGeometry:        addGeometry
+        ,setTokenSelection:  tokenSelectionUpdated
+        ,setCenter:          setCenter
+        ,setZoom:            setZoom
+        ,setEditMode:        setEditMode
+        ,setDrawType:        setDrawType
     };
 
     function insertMap() {
+        gmap = createGMap();
+        view = createView();
+
         var olMapDiv = document.getElementById('omap2');
         map = new ol.Map({
-            layers: layers,
+            layers: [],  // none as we are integrating the google map via external mechanism
             interactions: ol.interaction.defaults({
                 altShiftDragRotate: false,
                 dragPan: false,
@@ -65,6 +88,7 @@ function olmapFactory() {
         });
 
         view.on('change:center', function() {
+            //console.log('change:center', view.getCenter());
             var center = ol.proj.transform(view.getCenter(), 'EPSG:3857', 'EPSG:4326');
             gmap.setCenter(new google.maps.LatLng(center[1], center[0]));
         });
@@ -72,6 +96,7 @@ function olmapFactory() {
             //console.log("change:resolution", view.getZoom());
             gmap.setZoom(view.getZoom());
         });
+        // TODO handle size change.
 
         setCenter([-122.0, 36.83]);
         setZoom(10);
@@ -79,11 +104,193 @@ function olmapFactory() {
         olMapDiv.parentNode.removeChild(olMapDiv);
         gmap.controls[google.maps.ControlPosition.TOP_LEFT].push(olMapDiv);
 
-        prepareEditing();
+        createFeatureOverlay();
+        createModifyInteraction();
     }
 
+    function reinit() {
+        clearFeaturesOverlay();
+        if (map) {
+            map.getLayers().clear();
+        }
+    }
+
+    /**
+     * Adds a geometry to the map.
+     * @param geomId    ID of geometry
+     * @param geometry  GeoJSON object
+     */
+    function addGeometry(geomId, geometry) {
+        //console.warn("addGeometry geomId=", geomId, "geometry=", geometry);
+
+        var object = geometry;
+        if (!object.crs) {
+            // TODO perhaps force crs to always be stored.
+            // for now, just manually assigning out "well-knwon crs.
+            object.crs = {
+                type: 'name',
+                properties: {'name': 'EPSG:4326'}
+            };
+        }
+
+        var vectorSource = new ol.source.GeoJSON({
+            projection: 'EPSG:3857',
+            object: object
+        });
+
+        vectorLayer = new ol.layer.Vector({
+            source: vectorSource,
+            style:  styles.styleNormal
+        });
+        map.addLayer(vectorLayer);
+        vectorsByGeomId[geomId] = vectorLayer;
+    }
+
+    function tokenSelectionUpdated(selectedTokens) {
+        console.log("******* tokenSelectionUpdated=", selectedTokens);
+
+        if (editInfo.editingToken) {
+            endEditing();
+        }
+
+        tokenSelection = selectedTokens;
+        updateStylesForSelection();
+    }
+
+    function updateStylesForSelection() {
+        var selectedGeomIds = _.map(tokenSelection, "token_id");
+        console.log("selectedGeomIds=", selectedGeomIds);
+        _.each(vectorsByGeomId, function(vectorLayer, geomId) {
+            if (_.contains(selectedGeomIds, geomId)) {
+                vectorLayer.setStyle(styles.styleSelected);
+            }
+            else {
+                vectorLayer.setStyle(styles.styleNormal);
+            }
+        });
+    }
+
+    function setEditMode(editMode) {
+        console.log("******* setEditMode=", editMode);
+        if (editMode) {
+            if (editInfo.editingToken) {
+                return;
+            }
+            map.addInteraction(modify);
+
+            clearFeaturesOverlay();
+
+            // TODO for now not handling multiple selection for editing
+            if (tokenSelection.length == 1) {
+                startEditing(tokenSelection[0]);
+            }
+            else if (tokenSelection.length > 1) {
+                console.log("tokenSelectionUpdated: No multiple selection handled for editing");
+            }
+
+        }
+        else {
+            if (!editInfo.editingToken) {
+                return;
+            }
+            map.removeInteraction(modify);
+
+            endEditing();
+            //setDrawType(null); // no draw type
+
+            updateStylesForSelection();
+        }
+    }
+
+    function startEditing(token) {
+        if (!token.geometry) {
+            console.log("WARN: startEditing token_id=", token.token_id, " doesn't have geometry");
+            return;
+        }
+        var geomId = token.token_id;
+        var vectorLayer = vectorsByGeomId[geomId];
+        if (!vectorLayer) {
+            console.log("WARN: startEditing geomId=", geomId, "no such vector");
+            return;
+        }
+        console.log("startEditing geomId=", geomId, "vectorLayer=", vectorLayer);
+
+        if (editInfo.editingToken) {
+            endEditing();
+        }
+
+        editInfo.editingToken = token;
+
+        var rm = map.removeLayer(vectorLayer);
+        if ( rm !== vectorLayer) {
+            console.warn("removeLayer")
+        }
+
+        var source = vectorLayer.getSource();
+        var features = source.getFeatures();
+        addFeaturesToOverlay(features);
+    }
+
+    function addFeaturesToOverlay(features) {
+        console.warn("addFeaturesToOverlay", features);
+        for (var ii = 0; ii < features.length; ii++) {
+            var feature = features[ii];
+            featureOverlay.addFeature(feature);
+        }
+        console.log("---startEditing, overlayFeatures=", featureOverlay.getFeatures().getLength());
+    }
+
+    function endEditing() {
+        if (editInfo.editingToken) {
+            var token = editInfo.editingToken;
+            editInfo.editingToken = null;
+            var vectorLayer = updateLayer();
+            vectorsByGeomId[token.token_id] = vectorLayer;
+            map.addLayer(vectorLayer);
+
+            // update token.geometry and notify:
+            var features = vectorLayer.getSource().getFeatures();
+            var jsonFormat = new ol.format.GeoJSON();
+            var geometryString = jsonFormat.writeFeatures(features, {
+                dataProjection: 'EPSG:4326',
+                featureProjection: 'EPSG:3857'
+            });
+            token.geometry = angular.fromJson(geometryString);
+            console.warn("---endEditing token.geometry=", token.geometry);
+            $rootScope.$broadcast("tokenGeometryUpdated", token.token_id, token.geometry);
+        }
+        clearFeaturesOverlay();
+
+        // returns new layer with features in overlay
+        function updateLayer() {
+            var overlayFeatures = featureOverlay.getFeatures();
+            console.log("---updateLayer, overlayFeatures=", overlayFeatures.getLength());
+
+            var vectorSource = new ol.source.GeoJSON({
+                projection: 'EPSG:3857',
+                object: {
+                    type: 'FeatureCollection',
+                    crs: {
+                        type: 'name',
+                        properties: {'name': 'EPSG:4326'}
+                    }
+                    , features: []
+                }
+            });
+            for(var ii = 0; ii < overlayFeatures.getLength(); ii++) {
+                var feature = overlayFeatures.item(ii);
+                vectorSource.addFeature(feature);
+            }
+            vectorLayer = new ol.layer.Vector({
+                source: vectorSource,
+                style:  styles.styleNormal
+            });
+            return vectorLayer;
+        }
+    }
+    
     function setCenter(ll) {
-        view.setCenter(ol.proj.transform(ll, 'EPSG:4326', 'EPSG:3857'));
+        view.setCenter(transform(ll));
     }
 
     function setZoom(z) {
@@ -92,6 +299,10 @@ function olmapFactory() {
 
     ////////////////////////////////////////////////////////////////////////
     // private
+
+    function transform(ll) {
+        return ol.proj.transform(ll, 'EPSG:4326', 'EPSG:3857');
+    }
 
     function createGMap() {
         return new google.maps.Map(document.getElementById('gmap2'), {
@@ -107,72 +318,41 @@ function olmapFactory() {
 
     function createView() {
         return new ol.View({
-            // make sure the view doesn't go beyond the 22 zoom levels of Google Maps
-            maxZoom: 21
-        })
+            maxZoom: 21  // "make sure the view doesn't go beyond the 22 zoom levels of Google Maps"
+        });
     }
 
-    function createLayers() {
-        return [];
-        //var vector = new ol.layer.Vector({
-        //    source: new ol.source.GeoJSON({
-        //        url: 'olmap/data/countries.geojson',
-        //        projection: 'EPSG:3857'
-        //    }),
-        //    style: new ol.style.Style({
-        //        //fill: new ol.style.Fill({ color: 'rgba(255, 255, 255, 0.6)'}),
-        //        stroke: new ol.style.Stroke({ color: '#319FD3', width: 1})
-        //    })
-        //});
-        //return [vector]
-    }
-
-    function prepareEditing() {
-        // The features are not added to a regular vector layer/source,
+    function createFeatureOverlay() {
+        // "The features are not added to a regular vector layer/source,
         // but to a feature overlay which holds a collection of features.
         // This collection is passed to the modify and also the draw
-        // interaction, so that both can add or modify features.
+        // interaction, so that both can add or modify features."
         featureOverlay = new ol.FeatureOverlay({
-            style: new ol.style.Style({
-                fill: new ol.style.Fill({
-                    color: 'rgba(255, 255, 255, 0.2)'
-                }),
-                stroke: new ol.style.Stroke({
-                    color: '#ffcc33',
-                    width: 2
-                }),
-                image: new ol.style.Circle({
-                    radius: 7,
-                    fill: new ol.style.Fill({
-                        color: '#ffcc33'
-                    })
-                })
-            })
+            style: styles.styleOverlay
         });
         featureOverlay.setMap(map);
-
     }
 
-    function setEditMode(editMode) {
-        if (modify) {
-            map.removeInteraction(modify);
-            modify = null;
-        }
-        if (editMode) {
-            modify = new ol.interaction.Modify({
-                features: featureOverlay.getFeatures(),
-                // the SHIFT key must be pressed to delete vertices, so
-                // that new vertices can be drawn at the same position
-                // of existing vertices
-                deleteCondition: function(event) {
-                    return ol.events.condition.shiftKeyOnly(event) &&
-                        ol.events.condition.singleClick(event);
-                }
-            });
-            map.addInteraction(modify);
-        }
-        else {
-            setDrawType(null); // no draw type
+    function createModifyInteraction() {
+        modify = new ol.interaction.Modify({
+            features: featureOverlay.getFeatures(),
+            // "the SHIFT key must be pressed to delete vertices, so
+            // that new vertices can be drawn at the same position
+            // of existing vertices"
+            deleteCondition: function(event) {
+                return ol.events.condition.shiftKeyOnly(event) &&
+                    ol.events.condition.singleClick(event);
+            }
+        });
+    }
+
+    function clearFeaturesOverlay() {
+        if (featureOverlay) {
+            var overlayFeatures = featureOverlay.getFeatures();
+            if (overlayFeatures) {
+                console.log("!!!clearFeaturesOverlay, overlayFeatures=", overlayFeatures);
+                overlayFeatures.clear();
+            }
         }
     }
 
@@ -194,6 +374,30 @@ function olmapFactory() {
         }
     }
 
+    function getStyles() {
+        return {
+            styleNormal: new ol.style.Style({
+                fill: new ol.style.Fill({ color: 'rgba(255, 255, 255, 0.1)'}),
+                stroke: new ol.style.Stroke({color: '#319FD3', width: 2})
+            })
+
+            ,styleSelected: new ol.style.Style({
+                fill: new ol.style.Fill({color: 'rgba(255, 255, 255, 0.2)' }),
+                //fill: new ol.style.Fill({ color: 'rgba(255, 0, 0, 0.4)'}),
+                stroke: new ol.style.Stroke({color: '#ffcc33', width: 2})
+            })
+
+            ,styleOverlay: new ol.style.Style({
+                fill: new ol.style.Fill({ color: 'rgba(255, 255, 255, 0.3)' }),
+                stroke: new ol.style.Stroke({ color: '#ff0000', width: 2 }),
+                //stroke: new ol.style.Stroke({ color: '#ffcc33', width: 2 }),
+                image: new ol.style.Circle({
+                    radius: 7,
+                    fill: new ol.style.Fill({ color: '#ffcc33' })
+                })
+            })
+        };
+    }
 }
 
 })();